+++ src/all-components/main-menu/handleButtons.js
@@ -1,0 +1,9 @@
+import { clearScreen } from "../../all-utils/for-dom-calls";
+import { commenceCategorySelection } from "../starting-screen/categorySelection";
+import { addStartingScreenToDOM } from "../starting-screen/showLayout";
+
+export let handleButtons = () => {
+    clearScreen();
+    addStartingScreenToDOM();
+    commenceCategorySelection();
+}
